src/data_loader.py
"""
Data Loader Module
==================

Utilities for loading and acquiring NYC Taxi data.
All implementations are original, written from PySpark documentation.
"""

from pyspark.sql import SparkSession, DataFrame
from pyspark.sql.types import (
    StructType, StructField, IntegerType, DoubleType,
    StringType, TimestampType
)
from pyspark.sql.functions import col, input_file_name
import yaml
import os
from typing import Optional


def load_config(config_path: str = "config/spark_config.yaml") -> dict:
    """
    Load configuration from YAML file.
    
    Args:
        config_path: Path to configuration file
        
    Returns:
        Configuration dictionary
    """
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def create_spark_session(config: dict) -> SparkSession:
    """
    Create and configure Spark session for NYC Taxi ML pipeline.
    
    Original implementation following PySpark documentation.
    
    Args:
        config: Configuration dictionary
        
    Returns:
        Configured SparkSession
    """
    spark_config = config['spark']
    
    builder = SparkSession.builder \
        .appName(spark_config['app_name']) \
        .master(spark_config['master'])
    
    # Driver settings
    builder = builder.config("spark.driver.memory", spark_config['driver']['memory'])
    builder = builder.config("spark.driver.cores", spark_config['driver']['cores'])
    builder = builder.config("spark.driver.maxResultSize", spark_config['driver']['maxResultSize'])
    
    # Executor settings
    builder = builder.config("spark.executor.memory", spark_config['executor']['memory'])
    builder = builder.config("spark.executor.cores", spark_config['executor']['cores'])
    
    # SQL settings
    builder = builder.config("spark.sql.shuffle.partitions", 
                            spark_config['sql']['shuffle']['partitions'])
    builder = builder.config("spark.sql.adaptive.enabled", 
                            spark_config['sql']['adaptive']['enabled'])
    
    # Serialization
    builder = builder.config("spark.serializer", spark_config['serializer'])
    
    # Event logging
    if spark_config['eventLog']['enabled']:
        event_dir = spark_config['eventLog']['dir']
        os.makedirs(event_dir.replace("file://", ""), exist_ok=True)
        builder = builder.config("spark.eventLog.enabled", "true")
        builder = builder.config("spark.eventLog.dir", event_dir)
    
    spark = builder.getOrCreate()
    spark.sparkContext.setLogLevel("WARN")
    
    return spark


def get_nyc_taxi_schema() -> StructType:
    """
    Define explicit schema for NYC Taxi Trip Records.
    
    Original schema definition based on TLC data dictionary.
    18 columns as per 2023 yellow taxi format.
    
    Returns:
        StructType schema for NYC Taxi data
    """
    schema = StructType([
        StructField("VendorID", IntegerType(), True),
        StructField("tpep_pickup_datetime", TimestampType(), True),
        StructField("tpep_dropoff_datetime", TimestampType(), True),
        StructField("passenger_count", IntegerType(), True),
        StructField("trip_distance", DoubleType(), True),
        StructField("RatecodeID", IntegerType(), True),
        StructField("store_and_fwd_flag", StringType(), True),
        StructField("PULocationID", IntegerType(), True),
        StructField("DOLocationID", IntegerType(), True),
        StructField("payment_type", IntegerType(), True),
        StructField("fare_amount", DoubleType(), True),
        StructField("extra", DoubleType(), True),
        StructField("mta_tax", DoubleType(), True),
        StructField("tip_amount", DoubleType(), True),
        StructField("tolls_amount", DoubleType(), True),
        StructField("improvement_surcharge", DoubleType(), True),
        StructField("total_amount", DoubleType(), True),
        StructField("congestion_surcharge", DoubleType(), True)
    ])
    
    return schema


def load_raw_csv_rdd(spark: SparkSession, file_path: str) -> DataFrame:
    """
    Load NYC Taxi CSV using RDD-based parallelization.
    
    Original implementation for distributed CSV parsing.
    Uses RDD for raw parallelization before DataFrame conversion.
    
    Args:
        spark: SparkSession instance
        file_path: Path to CSV file(s) - supports wildcards
        
    Returns:
        DataFrame with NYC Taxi data
    """
    # Get schema
    schema = get_nyc_taxi_schema()
    
    # Load as RDD for parallel processing
    raw_rdd = spark.sparkContext.textFile(file_path)
    
    # Extract header
    header = raw_rdd.first()
    
    # Custom parsing function for NYC Taxi records
    def parse_taxi_record(line: str):
        """Parse single CSV line with validation."""
        if line == header:
            return None
            
        try:
            fields = line.split(',')
            
            # Validate field count
            if len(fields) != 18:
                return None
            
            # Basic validation - ensure critical fields are not empty
            if not fields[0] or not fields[1] or not fields[2]:
                return None
                
            return tuple(fields)
            
        except Exception:
            return None
    
    # Apply parsing in parallel
    data_rdd = raw_rdd.map(parse_taxi_record).filter(lambda x: x is not None)
    
    # Convert to DataFrame with schema
    df = spark.createDataFrame(data_rdd, schema=schema)
    
    return df


def load_raw_csv_dataframe(spark: SparkSession, file_path: str) -> DataFrame:
    """
    Load NYC Taxi CSV using DataFrame API (alternative method).
    
    Original implementation using Spark DataFrame reader.
    
    Args:
        spark: SparkSession instance
        file_path: Path to CSV file(s)
        
    Returns:
        DataFrame with NYC Taxi data
    """
    schema = get_nyc_taxi_schema()
    
    df = spark.read \
        .option("header", "true") \
        .schema(schema) \
        .csv(file_path)
    
    return df


def load_raw_parquet(spark: SparkSession, file_path: str) -> DataFrame:
    """
    Load NYC Taxi Parquet files without schema enforcement.
    
    Original implementation for Parquet format.
    Parquet files already contain schema, so we read them directly.
    
    Args:
        spark: SparkSession instance
        file_path: Path to Parquet file(s) - supports wildcards
        
    Returns:
        DataFrame with NYC Taxi data
    """
    # Read Parquet directly - schema is embedded in the file
    df = spark.read.parquet(file_path)
    
    return df


def add_bronze_metadata(df: DataFrame) -> DataFrame:
    """
    Add metadata columns for Bronze layer.
    
    Original implementation for data lineage tracking.
    
    Args:
        df: Input DataFrame
        
    Returns:
        DataFrame with metadata columns
    """
    from pyspark.sql.functions import year, month, lit
    from datetime import datetime
    
    # Cast integer columns to ensure consistency
    # NYC Taxi Parquet files may have different integer types
    df_casted = df \
        .withColumn("VendorID", col("VendorID").cast("integer")) \
        .withColumn("passenger_count", col("passenger_count").cast("integer")) \
        .withColumn("RatecodeID", col("RatecodeID").cast("integer")) \
        .withColumn("PULocationID", col("PULocationID").cast("integer")) \
        .withColumn("DOLocationID", col("DOLocationID").cast("integer")) \
        .withColumn("payment_type", col("payment_type").cast("integer"))
    
    df_with_metadata = df_casted \
        .withColumn("ingestion_timestamp", lit(datetime.now())) \
        .withColumn("source_file", input_file_name()) \
        .withColumn("data_year", year(col("tpep_pickup_datetime"))) \
        .withColumn("data_month", month(col("tpep_pickup_datetime")))
    
    return df_with_metadata


def save_to_parquet(df: DataFrame, output_path: str, 
                   partition_cols: Optional[list] = None,
                   mode: str = "overwrite") -> None:
    """
    Save DataFrame to Parquet format with optional partitioning.
    
    Original implementation for optimized storage.
    
    Args:
        df: DataFrame to save
        output_path: Output directory path
        partition_cols: Columns to partition by
        mode: Write mode (overwrite, append, etc.)
    """
    writer = df.write.mode(mode)
    
    if partition_cols:
        writer = writer.partitionBy(*partition_cols)
    
    writer.parquet(output_path)
    
    print(f"Data saved to: {output_path}")
    if partition_cols:
        print(f"Partitioned by: {', '.join(partition_cols)}")


def load_from_parquet(spark: SparkSession, input_path: str) -> DataFrame:
    """
    Load DataFrame from Parquet format.
    
    Args:
        spark: SparkSession instance
        input_path: Input directory path
        
    Returns:
        DataFrame loaded from Parquet
    """
    df = spark.read.parquet(input_path)
    return df


def validate_data_quality(df: DataFrame, config: dict) -> DataFrame:
    """
    Apply basic data quality validations for NYC Taxi data.
    
    Original implementation with NYC-specific business rules.
    
    Args:
        df: Input DataFrame
        config: Configuration dictionary with thresholds
        
    Returns:
        Validated DataFrame
    """
    features = config['project']['features']
    
    # Apply NYC Taxi-specific validation rules
    validated_df = df \
        .filter(col("fare_amount") >= features['min_fare']) \
        .filter(col("fare_amount") <= features['max_fare']) \
        .filter(col("trip_distance") >= features['min_distance']) \
        .filter(col("trip_distance") <= features['max_distance']) \
        .filter(col("passenger_count") >= features['min_passengers']) \
        .filter(col("passenger_count") <= features['max_passengers']) \
        .filter(col("tpep_pickup_datetime").isNotNull()) \
        .filter(col("tpep_dropoff_datetime").isNotNull()) \
        .filter(col("tpep_dropoff_datetime") > col("tpep_pickup_datetime")) \
        .filter(col("total_amount") > 0) \
        .filter(col("tip_amount") >= 0)
    
    return validated_df
