requirements.txt
pyspark==3.4.0
pandas==1.5.3
numpy==1.23.5
matplotlib==3.7.0
seaborn==0.12.2
jupyter==1.0.0
jupyterlab==3.6.0
mlflow==2.9.0
pyarrow==11.0.0
requests==2.31.0
pyyaml==6.0
